+++ Sources/MapboxSearch/PublicAPI/HistoryRecord.swift
@@ -25,6 +25,12 @@ public struct HistoryRecord: IndexableRecord, SearchResult, Codable, Hashable {
     /// Record's name
     public private(set) var name: String
     
+    /// Index in response from server.
+    public let serverIndex: Int?
+    
+    /// A point accuracy metric for the returned address.
+    public let accuracy: SearchResultAccuracy?
+    
     /**
         The feature name, as matched by the search algorithm.
         
@@ -92,6 +98,8 @@ public struct HistoryRecord: IndexableRecord, SearchResult, Codable, Hashable {
         id: String = UUID().uuidString,
         name: String,
         matchingName: String?,
+        serverIndex: Int?,
+        accuracy: SearchResultAccuracy?,
         coordinate: CLLocationCoordinate2D,
         timestamp: Date = Date(),
         historyType: HistoryRecord.HistoryType,
@@ -104,6 +112,8 @@ public struct HistoryRecord: IndexableRecord, SearchResult, Codable, Hashable {
         self.id = id
         self.name = name
         self.matchingName = matchingName
+        self.serverIndex = serverIndex
+        self.accuracy = accuracy
         self.coordinateCodable = .init(coordinate)
         self.timestamp = timestamp
         self.historyType = historyType
@@ -127,6 +137,8 @@ public struct HistoryRecord: IndexableRecord, SearchResult, Codable, Hashable {
         self.id = searchResult.id
         self.name = searchResult.name
         self.matchingName = searchResult.matchingName
+        self.serverIndex = searchResult.serverIndex
+        self.accuracy = searchResult.accuracy
         self.coordinateCodable = .init(searchResult.coordinate)
         self.timestamp = timestamp
         self.historyType = historyType