+++ Sources/MapboxSearch/PublicAPI/Search Results/SearchResult.swift
@@ -13,6 +13,12 @@ public protocol SearchResult {
     /// Icon name according to [Mapbox Maki icon set](https://github.com/mapbox/maki/)
     var iconName: String? { get }
     
+    /// Index in response from server.
+    var serverIndex: Int? { get }
+    
+    /// A point accuracy metric for the returned address.
+    var accuracy: SearchResultAccuracy? { get }
+    
     /// Type of SearchResult. Should be one of address or POI.
     var type: SearchResultType { get }