+++ Tests/MapboxSearchTests/Stubs&Models/SearchResultStub.swift
@@ -5,9 +5,11 @@ import CoreLocation
 class SearchResultStub: SearchResult {
     init(
         id: String,
+        accuracy: SearchResultAccuracy? = nil,
         categories: [String]? = nil,
         name: String,
         matchingName: String?,
+        serverIndex: Int?,
         iconName: String? = nil,
         resultType: SearchResultType,
         routablePoints: [RoutablePoint]? = nil,
@@ -17,9 +19,11 @@ class SearchResultStub: SearchResult {
         dataLayerIdentifier: String = "unit-test-stub"
     ) {
         self.id = id
+        self.accuracy = accuracy
         self.categories = categories
         self.name = name
         self.matchingName = matchingName
+        self.serverIndex = serverIndex
         self.iconName = iconName
         self.type = resultType
         self.routablePoints = routablePoints
@@ -32,9 +36,11 @@ class SearchResultStub: SearchResult {
     var dataLayerIdentifier: String
     
     var id: String
+    var accuracy: SearchResultAccuracy?
     var categories: [String]?
     var name: String
     var matchingName: String?
+    var serverIndex: Int?
     var iconName: String?
     var type: SearchResultType
     var routablePoints: [RoutablePoint]?